Create xor.c
The input layer is the first layer of the MLP, and it contains two neurons. The hidden layer is an intermediate layer between the input and output layers. In my architecture, it consists of two neurons. The output layer is the final layer of the MLP, and it contains one neuron.